requirements.txt
# Zoom Manager Dependencies
# This file is auto-generated from pyproject.toml
# Use: uv pip compile pyproject.toml -o requirements.txt
# Or just use: uv sync

# Core async & HTTP
aiohttp>=3.13.1
httpx>=0.24.0
asyncpg>=0.29.0

# Web framework
fastapi>=0.128.0
uvicorn>=0.40.0
pydantic>=2.0.0
pydantic-settings>=2.11.0
python-multipart>=0.0.6
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
greenlet>=3.2.4

# Task queue
celery>=5.6.1
redis>=5.0.0
celery-sqlalchemy-scheduler>=0.3.0
croniter>=2.0.1
pytz>=2024.1

# Security
bcrypt>=5.0.0
cryptography>=42.0.0
PyJWT>=2.8.0
python-ulid>=2.2.0

# Logging & Config
click>=8.0.0
loguru>=0.7.0
python-dotenv>=1.1.1

# Video processing
ffmpeg-python>=0.2.0

# Google/YouTube API
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# VK API
vk-api>=11.10.0

# AI/Transcription
fireworks-ai>=0.15.15
openai>=2.8.1

# Dev tools
ruff>=0.14.1
